Add sample codes for a HTTP command service
We sometimes want to trigger device actions from the cloud; that is,
from the Azure IoT Hub to EdgeX. The EdgeX Core Command Service
provides a comprehensive set of APIs to achieve this. However, in
some cases, you might prefer not to expose all APIs and therefore
require a finer-grained control over the APIs to be exposed. For
example, you might want to control the commands on specific devices
that can receive commands from outside of EdgeX, or to allow only
certain values for a specific command.  This commit adds
implementations of a sample HTTP Command Service to achieve such
control.

Signed-off-by: Jude Hung <jude@iotechsys.com>